Patient files (#9787)
