+++ public/locale/en.json
@@ -313,6 +313,7 @@
   "action_irreversible": "This action is irreversible",
   "actions": "Actions",
   "active": "Active",
+  "active_encounters": "Active Encounters",
   "active_files": "Active Files",
   "active_prescriptions": "Active Prescriptions",
   "add": "Add",
@@ -490,6 +491,7 @@
   },
   "board": "Board",
   "board_view": "Board View",
+  "book_a_new_appointment": "Book a new appointment",
   "book_an_appointment_with": "Book an appointment with",
   "book_appointment": "Book Appointment",
   "booked": "Booked",
@@ -668,10 +670,12 @@
   "covid_19_death_reporting_form_1": "Covid-19 Death Reporting : Form 1",
   "covid_details": "Covid Details",
   "create": "Create",
+  "create_a_new_encounter_to_get_started": "Create a new encounter to get started",
   "create_abha_address": "Create ABHA Address",
   "create_add_more": "Create & Add More",
   "create_asset": "Create Asset",
   "create_consultation": "Create Consultation",
+  "create_encounter": "Create Encounter",
   "create_facility": "Create a new facility",
   "create_new_abha_address": "Create New ABHA Address",
   "create_new_abha_profile": "Don't have an ABHA Number",
@@ -843,6 +847,7 @@
   "encounter_admit_sources__outp": "From outpatient department",
   "encounter_admit_sources__psych": "From psychiatric hospital",
   "encounter_admit_sources__rehab": "From rehabilitation facility",
+  "encounter_class": "Encounter Class",
   "encounter_class__amb": "Ambulatory",
   "encounter_class__completed": "Completed",
   "encounter_class__emer": "Emergency",
@@ -1011,6 +1016,7 @@
   "file_error__file_name": "Please give a name for all files!",
   "file_error__file_size": "Maximum size of files is 100 MB",
   "file_error__file_type": "Invalid file type \".{{extension}}\" Allowed types: {{allowedExtensions}}",
+  "file_error__mark_complete_failed": "Error while marking file upload as complete",
   "file_error__network": "Error Uploading File: Network Error",
   "file_error__single_file_name": "Please give a name for the file",
   "file_list_headings__consultation": "Consultation Files",
@@ -1277,6 +1283,7 @@
   "min_password_len_8": "Minimum password length 8",
   "min_time_bw_doses": "Min. time b/w doses",
   "minimize": "Minimize",
+  "missing_required_params_for_patient_verification": "Missing required parameters for patient verification",
   "mobile": "Mobile",
   "mobile_number": "Mobile Number",
   "mobile_number_different_from_aadhaar_mobile_number": "We have noticed that you have entered a mobile number that is different from the one linked to your Aadhaar. We will send an OTP to this number to link it with your Abha number.",
@@ -1334,6 +1341,7 @@
   "no_data_found": "No data found",
   "no_doctors_found": "No Doctors Found",
   "no_duplicate_facility": "You should not create duplicate facilities",
+  "no_encounters_found": "No encounters found",
   "no_facilities": "No Facilities found",
   "no_facilities_found": "No facilities found",
   "no_files_found": "No {{type}} files found",
@@ -1488,6 +1496,7 @@
   "patient_name": "Patient Name",
   "patient_name_uhid": "Patient Name/UHID",
   "patient_no": "OP/IP No",
+  "patient_not_found": "Patient Not Found",
   "patient_notes_thread__Doctors": "Doctor's Discussions",
   "patient_notes_thread__Nurses": "Nurse's Discussions",
   "patient_phone_number": "Patient Phone Number",
@@ -1622,6 +1631,8 @@
   "questionnaire_not_exist": "The questionnaire you tried to access does not exist.",
   "questionnaire_submission_failed": "Failed to submit questionnaire",
   "questionnaire_submitted_successfully": "Questionnaire submitted successfully",
+  "quick_actions": "Quick Actions",
+  "quick_actions_description": "Schedule an appointment or create a new encounter",
   "raise_consent_request": "Raise a consent request to fetch patient records over ABDM",
   "ration_card__APL": "APL",
   "ration_card__BPL": "BPL",
@@ -1740,6 +1751,7 @@
   "saving": "Saving...",
   "scan_asset_qr": "Scan Asset QR!",
   "schedule": "Schedule",
+  "schedule_an_appointment_or_create_a_new_encounter": "Schedule an appointment or create a new encounter",
   "schedule_appointment": "Schedule Appointment",
   "schedule_availability_created_successfully": "Availability created successfully",
   "schedule_availability_deleted_successfully": "Schedule availability deleted successfully",
@@ -1800,6 +1812,7 @@
   "select_facility_for_discharged_patients_warning": "Facility needs to be selected to view discharged patients.",
   "select_for_administration": "Select for Administration",
   "select_frequency": "Select frequency",
+  "select_gender": "Select gender",
   "select_groups": "Select Groups",
   "select_intent": "Select intent",
   "select_investigation": "Select Investigations (all investigations will be selected by default)",
@@ -1889,7 +1902,9 @@
   "staff_list": "Staff List",
   "staff_not_found": "Staff Not Found",
   "staff_username_not_found": "Staff username not found",
+  "start_a_new_clinical_encounter": "Start a new clinical encounter",
   "start_consultation": "Start Consultation",
+  "start_date": "Start Date",
   "start_datetime": "Start Date/Time",
   "start_dosage": "Start Dosage",
   "start_review": "Start Review",
@@ -2041,6 +2056,7 @@
   "upload": "Upload",
   "upload_an_image": "Upload an image",
   "upload_file": "Upload File",
+  "upload_files": "Upload Files",
   "upload_headings__consultation": "Upload New Consultation File",
   "upload_headings__patient": "Upload New Patient File",
   "upload_headings__sample_report": "Upload Sample Report",
@@ -2108,6 +2124,7 @@
   "view": "View",
   "view_abdm_records": "View ABDM Records",
   "view_all_details": "View All Details",
+  "view_and_manage_patient_encounters": "View and manage patient encounters",
   "view_asset": "View Assets",
   "view_cns": "View CNS",
   "view_consultation": "View Latest Encounter",
