+++ src/pages/Patients/VerifyPatient.tsx
@@ -2,6 +2,7 @@ import { useMutation, useQuery } from "@tanstack/react-query";
 import { AlertCircle, CalendarIcon } from "lucide-react";
 import { Link, useQueryParams } from "raviger";
 import { useEffect } from "react";
+import { useTranslation } from "react-i18next";
 import { toast } from "sonner";
 
 import CareIcon from "@/CAREUI/icons/CareIcon";
@@ -28,6 +29,7 @@ import { formatPatientAge } from "@/Utils/utils";
 import { Encounter } from "@/types/emr/encounter";
 
 export default function VerifyPatient(props: { facilityId: string }) {
+  const { t } = useTranslation();
   const [qParams] = useQueryParams();
   const { phone_number, year_of_birth, partial_id } = qParams;
 
@@ -69,7 +71,7 @@ export default function VerifyPatient(props: { facilityId: string }) {
         <Alert variant="destructive">
           <AlertCircle className="h-4 w-4" />
           <AlertDescription>
-            Missing required parameters for patient verification
+            {t("missing_required_params_for_patient_verification")}
           </AlertDescription>
         </Alert>
       ) : patientData ? (
@@ -95,10 +97,10 @@ export default function VerifyPatient(props: { facilityId: string }) {
                       <h3 className="text-sm font-medium text-gray-600">
                         {formatPatientAge(patientData, true)},{"  "}
                         <span className="capitalize">
-                          {patientData.gender.replace("_", " ")},{"  "}
+                          {patientData.gender.replace("_", " ")}
                         </span>
                         {patientData.blood_group &&
-                          patientData.blood_group.replace("_", " ")}
+                          ", " + patientData.blood_group.replace("_", " ")}
                       </h3>
                     </div>
                   </div>
@@ -109,9 +111,9 @@ export default function VerifyPatient(props: { facilityId: string }) {
 
           <Card>
             <CardHeader>
-              <CardTitle>Quick Actions</CardTitle>
+              <CardTitle>{t("quick_actions")}</CardTitle>
               <CardDescription>
-                Schedule an appointment or create a new encounter
+                {t("quick_actions_description")}
               </CardDescription>
             </CardHeader>
             <CardContent className="grid gap-4 sm:grid-cols-1 md:grid-cols-2">
@@ -131,10 +133,10 @@ export default function VerifyPatient(props: { facilityId: string }) {
                     </div>
                     <div className="flex flex-col items-start gap-0.5">
                       <span className="text-base md:text-lg font-semibold text-gray-800 group-hover:text-primary transition-colors line-clamp-1">
-                        Schedule Appointment
+                        {t("schedule_appointment")}
                       </span>
                       <span className="text-xs md:text-sm text-gray-500 line-clamp-1">
-                        Book a new appointment
+                        {t("book_a_new_appointment")}
                       </span>
                     </div>
                     <CareIcon
@@ -166,10 +168,10 @@ export default function VerifyPatient(props: { facilityId: string }) {
                         </div>
                         <div className="flex flex-col items-start gap-0.5">
                           <span className="text-base md:text-lg font-semibold text-gray-800 group-hover:text-primary transition-colors line-clamp-1">
-                            Create Encounter
+                            {t("create_encounter")}
                           </span>
                           <span className="text-xs md:text-sm text-gray-500 line-clamp-1">
-                            Start a new clinical encounter
+                            {t("start_a_new_clinical_encounter")}
                           </span>
                         </div>
                         <CareIcon
@@ -186,9 +188,9 @@ export default function VerifyPatient(props: { facilityId: string }) {
 
           <Card>
             <CardHeader className="pb-2">
-              <CardTitle>Active Encounters</CardTitle>
+              <CardTitle>{t("active_encounters")}</CardTitle>
               <CardDescription>
-                View and manage patient encounters
+                {t("view_and_manage_patient_encounters")}
               </CardDescription>
             </CardHeader>
             <CardContent className="flex flex-col gap-3 pt-2">
@@ -207,10 +209,10 @@ export default function VerifyPatient(props: { facilityId: string }) {
                     />
                   </div>
                   <h3 className="text-base md:text-lg font-semibold mb-1">
-                    No encounters found
+                    {t("no_encounters_found")}
                   </h3>
                   <p className="text-xs md:text-sm text-muted-foreground">
-                    Create a new encounter to get started
+                    {t("create_a_new_encounter_to_get_started")}
                   </p>
                 </div>
               )}
